interim test update - use sub objects for group
